Improve HUD stuff config and add abstract classes for them to implement.
Currently only DungeonMap is done as a prototype.

The mess of setting and getting values from config can be avoided by passing it only once via MutableInt/MutableFloat on the constructor, but then every use of x/y/scale turns into an even worse unboxing mess. If there was a balance between the 2 it would be perfect.